PSHosts Variable Addition
Added the PSHosts variable to access the systems listview items and
return each column as a PSObject for use by the psf console and
powershell scripts.
Added hosttest.ps1 to show the usage of the new $PSHosts variable.
Added a comma between the IP address entries. (Not sure if I want to
split these out... )